Fix updating the counter for different rules
There is a bug on updating counter for all rules. If the smaller rule is reached, also the bigger rules were updating too. 
For example if we have PerHour = 1 and PerDay = 5. If you reach the hourly limit and continue making calls (6 for example) the second rule PerDay is also updated. And after one hour when you can make calls according the first rule you will get an error that the daily limit is reached